+++ SimPEG/electromagnetics/static/spontaneous_potential/__init__.py
@@ -1,4 +1,2 @@
-from .simulation import Problem_CC, BaseSPProblem, Problem_CC_Jstore, SurveySP_store
+from .simulation import Simulation3DCellCentered, Survey
 from . import sources
-from . import utils
-from SimPEG.EM.Static.DC import Rx, Survey
